+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # ember-polaris Changelog
 
+### Unreleased
+- [#144](https://github.com/smile-io/ember-polaris/pull/144) [UPDATE] Shopify Polaris `v2.2.0`
+  - Add `new` status to the Badge component
+
 ### v1.6.1 (June 28, 2018)
 
 - [#141](https://github.com/smile-io/ember-polaris/pull/141) [FEATURE] Add `polaris-data-table` component.