+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 - [#144](https://github.com/smile-io/ember-polaris/pull/144) [UPDATE] Shopify Polaris `v2.2.0`
   - [#146](https://github.com/smile-io/ember-polaris/pull/146) Add `new` status to the Badge component
+  - [#152](https://github.com/smile-io/ember-polaris/pull/152) Accept card title as a string or a component
 
 ### v1.6.1 (June 28, 2018)