+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ### Unreleased
 - [#144](https://github.com/smile-io/ember-polaris/pull/144) [UPDATE] Shopify Polaris `v2.2.0`
-- [#146](https://github.com/smile-io/ember-polaris/pull/146) Add `new` status to the Badge component
+- [#146](https://github.com/smile-io/ember-polaris/pull/146) [ENHANCEMENT] Add `new` status to the Badge component
 - [#148](https://github.com/smile-io/ember-polaris/pull/148) [FEATURE] Add `polaris-label` and `polaris-labelled` components.
 - [#151](https://github.com/smile-io/ember-polaris/pull/151) [FIX] Add missing `hidden` class to `polaris-label`.
 - [#152](https://github.com/smile-io/ember-polaris/pull/152) Accept card title as a string or a component
 - [#154](https://github.com/smile-io/ember-polaris/pull/154) [FEATURE] Add `polaris-range-slider` component.
+- [#152](https://github.com/smile-io/ember-polaris/pull/152) [ENHANCEMENT] Accept card title as a string or a component
+- [#155](https://github.com/smile-io/ember-polaris/pull/155) [ENHANCEMENT] Add support for extra large dropzone
 
 ### v1.6.1 (June 28, 2018)